pyproject.toml
[tool.poetry]
name = "vaultpay"
version = "0.1.0"
description = "Fintech API backend for VaultPay (FastAPI + SQLAlchemy + Postgres)"
authors = ["Abhinav Raj <rajabhinav1001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.13"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.11.7"}
alembic = "^1.16.4"
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
requests = "^2.32.5"
bcrypt = "^4.3.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
